Get Jest and Mocha tests running together - introduce jestBabelTransform to graphiql - rename all jest tests to .spec.js - remove mocha imports from each test in graphql-language-service packages - introduce jest & mocha to eslint
