+++ babel.config.js
@@ -5,6 +5,7 @@ module.exports = {
     require.resolve('@babel/preset-react'),
   ],
   plugins: [
-    require.resolve('@babel/plugin-proposal-class-properties')
+    require.resolve('@babel/plugin-proposal-class-properties'),
+    // require.resolve('@babel/plugin-transform-runtime'),
   ],
 };
