+++ packages/codemirror-graphql/package.json
@@ -28,16 +28,10 @@
     "README.md",
     "LICENSE"
   ],
-  "options": {
-    "mocha": "--full-trace --require resources/mocha-bootload src/**/__tests__/**/*-test.js",
-    "mocha_tdd": "--full-trace --watch --require resources/mocha-bootload src/**/__tests__/**/*-test.js"
-  },
   "scripts": {
-    "test": "babel-node ./node_modules/.bin/_mocha $npm_package_options_mocha",
-    "testwatch": "babel-node --inspect ./node_modules/.bin/_mocha $npm_package_options_mocha_tdd",
     "lint": "eslint src",
     "check": "flow check",
-    "build": "yarn run build-js && yarn run build-flow",
+    "build": "babel src --root-mode upward --ignore __tests__ --out-dir . && yarn run build-flow",
     "build-js": "node ../../resources/buildJs.js",
     "build-flow": "node ../../resources/buildFlow.js",
     "watch": "babel --optional runtime resources/watch.js | node",
