+++ packages/codemirror-graphql/src/__tests__/hint-test.js
@@ -8,7 +8,7 @@
  */
 
 import { expect } from 'chai';
-import { describe, it } from 'mocha';
+// 
 import CodeMirror from 'codemirror';
 import 'codemirror/addon/hint/show-hint';
 import { isCompositeType } from 'graphql';
