+++ packages/codemirror-graphql/src/__tests__/lint-test.js
@@ -8,7 +8,6 @@
  */
 
 import { expect } from 'chai';
-import { describe, it } from 'mocha';
 import CodeMirror from 'codemirror';
 import 'codemirror/addon/lint/lint';
 import '../lint';
