+++ packages/codemirror-graphql/src/utils/__tests__/jsonParse-test.js
@@ -8,7 +8,7 @@
  */
 
 import { expect } from 'chai';
-import { describe, it } from 'mocha';
+
 import jsonParse from '../jsonParse';
 
 describe('jsonParse', () => {
