+++ packages/graphiql/package.json
@@ -33,11 +33,10 @@
     "setupFiles": [
       "<rootDir>/resources/enzyme.config.js"
     ],
+    "transform": {
+      "^.+\\.jsx?$": "./resources/jestBabelTransform"
+    },
     "testEnvironment": "jsdom",
-    "testMatch": [
-      "**/__tests__/**/*-test.js",
-      "**/?(*.)+(spec|test).js?(x)"
-    ],
     "testPathIgnorePatterns": [
       "\\\\node_modules\\\\",
       "dist"
@@ -51,7 +50,7 @@
     "dev": "babel-node test/server.js",
     "prepublish": ". ./resources/prepublish.sh",
     "preversion": ". ./resources/checkgit.sh && npm test",
-    "test": "jest"
+    "test": "jest --no-cache spec"
   },
   "lint-staged": {
     "*.js": [
