+++ packages/graphiql/resources/enzyme.config.js
@@ -1,5 +1,12 @@
+// for regeneratorRuntime
+require('@babel/polyfill', {
+  rootMode: 'upward'
+})
+
+
 import { configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-15';
+
 
 configure({ adapter: new Adapter() });
