+++ packages/graphiql/src/components/DocExplorer/__tests__/FieldDoc.spec.js
