+++ packages/graphiql/src/components/DocExplorer/__tests__/TypeDoc-.spec.js
