+++ packages/graphiql/src/components/DocExplorer/__tests__/TypeLink.spec.js
