+++ packages/graphiql/src/components/__tests__/DocExplorer.spec.js
