+++ packages/graphiql/src/components/__tests__/GraphiQL.spec.js
