+++ packages/graphiql/src/utility/__tests__/getQueryFacts.spec.js
