+++ packages/graphiql/src/utility/__tests__/mergeAst.spec.js
