+++ packages/graphiql/src/utility/__tests__/normalizeWhitespace.spec.js
