+++ packages/graphql-language-service-interface/src/__tests__/getOutline-test.js
@@ -9,7 +9,7 @@
  */
 
 import { expect } from 'chai';
-import { describe, it } from 'mocha';
+
 import { getOutline } from '../getOutline';
 
 describe('getOutline()', () => {
