+++ resources/pretty.js
@@ -14,7 +14,7 @@ const INVERSE = '\x1b[7m';
 const RESET = '\x1b[0m';
 const YELLOW = '\x1b[33m';
 
-const options = ['--trailing-comma=all'];
+const options = ['--trailing-comma=es5'];
 const glob = '{packages/*/{resources,src},resources,src}/**/*.js';
 const root = join(__dirname, '..');
 const executable = join(root, 'node_modules', '.bin', 'prettier');
